Changed data type

+++ backend/Fung.COMMON/Entities/LevelIndicatorTransactions.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public double Height { get; set; }
-        public decimal Temperature { get; set; }
+        public double Temperature { get; set; }
         public double Volume { get; set; }
         public double Density { get; set; }
         public DateTime TransactionTime { get; set; }
